requirements.txt
asgiref==3.5.2
autoflake==1.4
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django==4.0.5
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docformatter==1.4
Pillow==9.1.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pyformat==0.7
PyJWT==2.4.0
python-decouple==3.6
python-slugify==6.1.2
pytz==2022.1
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
unify==0.5
untokenize==0.1.1